cvs/monitors/cluster-mon/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pydantic==2.5.3
pydantic-settings==2.1.0
parallel-ssh==2.12.0
paramiko==3.4.0
scp==0.14.5
redis==5.0.1
influxdb-client==1.39.0
python-multipart==0.0.6
pyyaml==6.0.1
asyncio==3.4.3
aiofiles==23.2.1
python-dateutil==2.8.2
requests==2.31.0
